requirements.txt
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==0.9.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
